pyproject.toml
[tool.poetry]
name = "recipebook"
version = "0.1.1"
description = "I'm writing a cookbook to learn Python, Postgresql, and other potential tools."
authors = ["Mike Martino <mikemartino86@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/mikemartino/recipebook"

[tool.poetry.urls]
"PyPI" = "https://pypi.org/project/recipebook/"

[tool.poetry.dependencies]
python = "^3.7"
psycopg2-binary = "^2.8.5"
pydantic = "^1.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
save-recipe = "script.save_recipe:main"
get-recipe = "script.get_recipe:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

